Added loading overlay
- Added control "LoadingOverlay". Can be toggled with LoadingStarted()
and LoadingFinished()
- Added loading overlay to login page (closes #48)
- Small bugfix for a null reference exception in Task.cs

+++ SparklrWP/Utils/Task.cs
@@ -10,7 +10,8 @@ namespace SparklrWP.Utils
         public Task()
         {
 #if DEBUG
-            ScheduledActionService.Remove(periodicTaskName);
+            if (ScheduledActionService.Find(periodicTaskName) != null)
+                ScheduledActionService.Remove(periodicTaskName);
 #endif
             periodicTask = ScheduledActionService.Find(periodicTaskName) as PeriodicTask;
             if (periodicTask == null)
